# src/main/resources/documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo Ecommerce API",
    "description": "API documentation for managing products, sales and users",
    "contact": {
      "name": "Diego Pimenta",
      "url": "https://github.com/Diego-Pimenta",
      "email": "diegopimenta888@gmail.com"
    },
    "license": {
      "name": "Apache License",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/sales/{id}": {
      "get": {
        "tags": [
          "Sale",
          "Get"
        ],
        "summary": "Retrieve a sale by id",
        "description": "Get a sale object by specifying its id",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sale successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Sale not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      },
      "put": {
        "tags": [
          "Sale",
          "Put"
        ],
        "summary": "Update a sale by id",
        "description": "Modify a sale object with body content by its id",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sale successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Sale data not achievable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Delete",
          "Sale"
        ],
        "summary": "Delete a sale by id",
        "description": "Remove a sale object by specifying its id",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sale successfully deleted"
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Sale not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      }
    },
    "/v1/products/{id}": {
      "get": {
        "tags": [
          "Get",
          "Product Stock"
        ],
        "summary": "Retrieve a product in stock by id",
        "description": "Get a product object by specifying its id",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product successfully retrieved from stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStockResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Product not found in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      },
      "put": {
        "tags": [
          "Product Stock",
          "Put"
        ],
        "summary": "Update a product in stock by id",
        "description": "Modify a product object with body content by its id",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductStockRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product successfully updated in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStockResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Product not found in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Delete",
          "Product Stock"
        ],
        "summary": "Delete a product in stock by id",
        "description": "Remove a product object by specifying its id",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product successfully deleted from stock"
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Product not found in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      }
    },
...
    "/v1/users/status/{cpf}": {
      "patch": {
        "tags": [
          "User",
          "Patch"
        ],
        "summary": "Update an user status by cpf",
        "description": "Modify an user object with body content by its cpf",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActiveStatusRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      }
    },
...
    "/v1/sales/status/{id}": {
      "patch": {
        "tags": [
          "Sale",
          "Patch"
        ],
        "summary": "Update a sale status by id",
        "description": "Modify a sale object with body content by its id",
        "operationId": "updateStatus_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleStatusRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sale successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Sale data not achievable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      }
    },
    "/v1/products/status/{id}": {
      "patch": {
        "tags": [
          "Patch",
          "Product Stock"
        ],
        "summary": "Update a product status in stock by id",
        "description": "Modify a product object with body content by its id",
        "operationId": "updateStatus_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActiveStatusRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product successfully updated in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStockResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation with data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          },
          "404": {
            "description": "Product not found in stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardError"
                }
              }
            }
          }
        },
        "security": [
          {
            "security": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CLIENT",
              "ADMIN"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "phone_number": {
            "type": "string"
          }
        }
      },
      "StandardError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ProductStockResponseDto": {
        "required": [
          "@class"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "unit_price": {
            "type": "number"
          },
          "@class": {
            "type": "string"
          }
        },
        "discriminator": {
          "propertyName": "@class"
        }
      },
      "SaleItemResponseDto": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ProductStockResponseDto"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "sub_total": {
            "type": "number"
          }
        }
      },
      "SaleResponseDto": {
        "required": [
          "@class"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "moment": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING_PAYMENT",
              "PAID",
              "SHIPPED",
              "DONE",
              "CANCELLED"
            ]
          },
          "customer": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "@class": {
            "type": "string"
          }
        },
        "discriminator": {
          "propertyName": "@class"
        }
      },
      "SaleItemRequestDto": {
        "required": [
          "product_id",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "product_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SaleRequestDto": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemRequestDto"
            }
          }
        }
      },
      "UpdateProductStockRequestDto": {
        "required": [
          "active",
          "category",
          "name",
          "quantity",
          "unit_price"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "quantity": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "unit_price": {
            "maximum": 999999,
            "minimum": 0,
            "type": "number"
          }
        }
      },
...
      "UpdateActiveStatusRequestDto": {
        "required": [
          "active"
        ],
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateSaleStatusRequestDto": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "WAITING_PAYMENT",
              "PAID",
              "SHIPPED",
              "DONE",
              "CANCELLED"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "security": {
        "type": "http",
        "description": "Authentication required with bearer token",
        "name": "security",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}